cmd/utils: set bootnodes via --chainspec.parity
This is ugly. I don't like it.
It reads and Marshals the file twice.

Ideally params.ChainConfig should contain field
Bootnodes and we can get rid of another 8 global variables.

But that requires reading and editing the set-configuration
logic around node and p2p setup and that's just more
complex than this is worth right now.

But for the record, I don't like it.
Not one bit.